Refactor minimax to negamax
The negamax algorithm allows for code with less duplication, but is
arguably more difficult to follow.  I'm undecided on whether I like
minimax or negamax better.

I also refactored the algorithm to using the state of the board to
determine whether to be in min mode or max mode.  I like this better
because I don't have to worry about making sure the isMaxPlayer flag is
in sync with the state of the board.