Decouple core board position names from UI
This ended up touching a lot of different areas for a couple of reasons.
First, there were a bunch of tests that had to be rewritten to use the
new labels.  Second, ConsolePlayer class could no longer be used
generically between the game versions because the position mapping is
different for each game variation.  I split it into two classes.  I also
added tests for these two classes.